Add workaround for libcurl/mbedtls async bug.
When compiled with mbedtls, libcurl versions prior to <7.530 has a bug
with its async interface, which manifests as libcurl informing the app
to close the socket and retrieve results immediately after the socket
is written to (see https://github.com/curl/curl/commit/b993d2cc).

This change adds a hack to detect this at run time and if found
effectively falls back to synchronous behaviour, with the resulting
performance degradation.  The hack is on by default but can be disabled
with -DNO_LIBCURL_BUG_WORKAROUND.