Add a timer to flush the debug log.
When debug logging is enabled debug-level messages are not flushed,
presumably to keep the overhead down.  This means that debug messages
can end up not being shown until either a more severe error happens or
the server gets shut down.  This adds a timer that flushes the log
every ten seconds.  (This is set up after logging_init since the log
may be needed during the event loop setup.)