Use proxy remote name resolution if possible.
Many of the protocols supported by libcurl support remote name
resolution.  When those are in use, do not attempt local resolution of
dns.google.com and instead leave it to the proxy to resolve.  This may
result in better performance if the proxy gets a closer address for
the server.