Use FindGTest to search for googletest
On Ubuntu 19.04, the previous way of doing things caused build errors
due to the lack of -lpthread. While explicitly adding
find_package(Threads) worked as well, this seems like a better option.

This also fixes an issue on Debian Jessie where the build would fail if
gtest.h exists but libgtest.a does not. Jessie only includes the
source files with the libgtest-dev package, and requires manually
building libgtest.a after installation.

Signed-off-by: Matt Merhar <mattmerhar@protonmail.com>