Remove the EDNS (-e) option.
Google supported this via a HTTP GET argument on it's JSON protocol but we now require the binary
RFC-8484 format which is both safer and simpler.

Note that EDNS client-subnet (https://tools.ietf.org/html/rfc7871) may still be supported by DNS providers, but it will be supported as a DNS extension.